requirements.txt
streamlit
opencv-python-headless
torch
torchvision
pandas
pillow
numpy
matplotlib
pyyaml
requests
seaborn
tqdm
jupyter
scikit-learn
ultralytics